refactor: move key vertices finder algorithm to separate class
